test(integration): reorganize and simplify the integration tests
