+++ integration/main_test.go
@@ -1,0 +1,11 @@
+//go:build integration
+
+package integration
+
+import (
+	"os"
+)
+
+func init() {
+	os.Setenv("TESTCONTAINERS_RYUK_DISABLED", "true")
+}
